refactor: 💡 create layout examples-starter
- remove layout/demo
